// React/app/GerAdv_TS/GUTMatriz/Interfaces/interface.GUTMatriz.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IGUTMatriz {
// 202501251
    id: number;
 
	guttipo: number,
	descricao: string,
	valor: number,
}

export interface IGUTMatrizFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IGUTMatrizIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}